check user id before deleting

+++ renderer/src/main/scala/com/olegych/scastie/PastesActor.scala
@@ -16,44 +16,49 @@ class PastesActor(pastesContainer: PastesContainer, progressActor: ActorRef) ext
   val failures = context.actorOf(Props[FailuresActor], "failures")
 
   def receive = LoggingReceive {
-    case AddPaste(content) =>
+    case AddPaste(content, uid) =>
       val id = nextPasteId
-      val paste = Paste(id = id, content = Option(content), output = Option("Processing..."))
+      val paste = Paste(id = id, content = Option(content), output = Option("Processing..."), uid = Some(uid))
       renderer ! paste
       sender ! paste
       writePaste(paste)
     case GetPaste(id) =>
       sender ! readPaste(id)
-    case DeletePaste(id) =>
-      deletePaste(id)
-      sender ! Paste(id, None, None)
-    case paste@Paste(id, content, output) =>
+    case DeletePaste(id, uid) =>
+      sender ! deletePaste(id, uid)
+    case paste: Paste =>
       writePaste(paste)
   }
 
   def writePaste(paste: Paste) {
     val pasteDir = pastesContainer.paste(paste.id)
     val contentChanged = readPaste(paste.id).content != paste.content
-    pasteDir.writeFile(pasteDir.pasteFile, paste.content)
+    pasteDir.pasteFile.write(paste.content)
+    pasteDir.uidFile.write(paste.uid)
     progressActor ! PasteProgress(paste.id, contentChanged, paste.output.getOrElse(""))
-    pasteDir.writeFile(pasteDir.outputFile, paste.output, truncate = false)
+    pasteDir.outputFile.write(paste.output, truncate = false)
   }
 
   def readPaste(id: Long) = {
     val paste = pastesContainer.paste(id)
     if (paste.pasteFile.exists()) {
-      import scalax.io.Resource._
-      Paste(id = id, content = Option(fromFile(paste.pasteFile).string),
-        output = Option(fromFile(paste.outputFile).string))
+      Paste(id = id, content = Option(paste.pasteFile.read),
+        output = Option(paste.outputFile.read), uid = Option(paste.uidFile.read))
     } else {
-      Paste(id = id, content = None, output = Option("Not found"))
+      Paste(id = id, content = None, output = Option("Not found"), uid = None)
     }
   }
 
-  def deletePaste(id: Long) {
-    val paste = pastesContainer.paste(id)
-    paste.outputFile.delete()
-    paste.pasteFile.delete()
+  def deletePaste(id: Long, uid: String) = {
+    val storedPaste = readPaste(id)
+    if (storedPaste.uid == Some(uid)) {
+      val paste = pastesContainer.paste(id)
+      paste.outputFile.delete()
+      paste.pasteFile.delete()
+      readPaste(id)
+    } else {
+      storedPaste
+    }
   }
 
   def nextPasteId = pastesContainer.lastPasteId.incrementAndGet()
@@ -63,13 +68,14 @@ object PastesActor {
 
   sealed trait PasteMessage
 
-  case class AddPaste(content: String) extends PasteMessage
+  case class AddPaste(content: String, uid: String) extends PasteMessage
 
   case class GetPaste(id: Long) extends PasteMessage
 
-  case class DeletePaste(id: Long) extends PasteMessage
+  case class DeletePaste(id: Long, uid: String) extends PasteMessage
 
-  case class Paste(id: Long, content: Option[String], output: Option[String]) extends PasteMessage
+  case class Paste(id: Long, content: Option[String], output: Option[String], uid: Option[String])
+      extends PasteMessage
 
   case class PasteProgress(id: Long, contentChanged: Boolean, output: String)
